+++ packages/koop-core/lib/datasets/model.js
@@ -24,15 +24,15 @@ Dataset.prototype.deleteFromCache = function (key, callback) {
 // Metadata Operations
 
 Dataset.prototype.getFromCatalog = function (key, options, callback) {
-  this.cache.catalog.retrieve(key, options, callback)
+  this.cache.catalogRetrieve(key, options, callback)
 }
 
 Dataset.prototype.insertIntoCatalog = function (key, metadata, callback) {
-  this.cache.catalog.insert(key, metadata, callback)
+  this.cache.catalogInsert(key, metadata, {}, callback)
 }
 
 Dataset.prototype.deleteFromCatalog = function (key, callback) {
-  this.cache.catalog.delete(key, callback)
+  this.cache.catalogDelete(key, callback)
 }
 
 module.exports = Dataset