+++ packages/koop-core/CHANGELOG.md
@@ -2,7 +2,7 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-## Unreleased
+## [5.0.0] - 2022-08-05
 ### Changed
 * Pull in updated cache-memory plugin
 * Remove support for Node 10
@@ -971,6 +971,7 @@ Koop is now just a node module that exposes an express middleware app with hooks
   - koop-server is no more; all central code is in the koop project
   - to use Koop you must use it as middleware in an app that boots up an http server
 
+[5.0.0]: https://github.com/koopjs/koop-core/compare/v4.2.4...v5.0.0
 [4.2.4]: https://github.com/koopjs/koop-core/compare/v4.2.3...v4.2.4
 [4.2.3]: https://github.com/koopjs/koop-core/compare/v4.2.2...v4.2.3
 [4.2.2]: https://github.com/koopjs/koop-core/compare/v4.2.1...v4.2.2